Working pages added
I got the cookie function to work. Right now, here's how it goes:
The user inputs a username/pass on createuser.php, which stores each as
a cookie lasting 30secs. The cookies get tested on loggedin.php for
whether the de-hashed, username-attached password matches the one in the
cookie. If it does, it echoes "successful login." If not, it redirects
to "notlogged.html," which gives an error and a link back to
createuser.php. This all requires PasswordHash.php to be in the same
folder, and it obviously requires passwords.txt in html_data.